Add files via upload
The Restorative Coffee Yemen Initiative aims to rejuvenate Yemen's historic coffee industry while addressing the urgent healthcare needs of millions affected by the civil war. Through the Mocha, Medicine, Mitigation (MMM) project, we seek to create a sustainable network of coffee mills that not only boost the local economy but also provide essential healthcare services to farmers and their families. By fostering healthier communities and supporting small growers, this initiative aspires to lay the groundwork for lasting peace and prosperity in Yemen. This repository will track the project's development, implementation, and impact, offering investors and stakeholders a transparent view of our efforts to restore both the economy and the well-being of Yemeni citizens.